refactor: decompose handwritten JSDoc coverage into focused modules
